+++ src/TinyGsmClientSIM7000.h
@@ -631,11 +631,11 @@ class TinyGsmSim7000
       // DBG("### Got Data:", len, "on", mux);
       return true;
     } else if (data.endsWith(GF("CLOSED\r\n"))) {
-      int8_t nl   = TinyGsmMax(0,
-                               data.lastIndexOf(String(GFP(ModemConfig::GSM_NL)),
-                                                data.length() - 8));
-      int8_t coma = data.indexOf(',', nl + 2);
-      int8_t mux  = data.substring(nl + 2, coma).toInt();
+      int nl   = TinyGsmMax(0,
+                            data.lastIndexOf(String(GFP(ModemConfig::GSM_NL)),
+                                             data.length() - 8));
+      int coma = data.indexOf(',', nl + 2);
+      int mux  = data.substring(nl + 2, coma).toInt();
       if (mux >= 0 && mux < TcpConfig::kMuxCount && sockets[mux]) {
         sockets[mux]->sock_connected = false;
       }
